requirements_dev.txt
alabaster==0.7.12
amqp==2.5.1
aniso8601==8.0.0
appdirs==1.4.3
argh==0.26.2
attrs==19.3.0
autoprotocol==6.1.0
Babel==2.7.0
backcall==0.1.0
billiard==3.6.1.0
bleach==3.1.0
bump2version==0.5.11
cached-property==1.5.1
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==4.5.4
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.8
Flask==1.1.1
Flask-RESTful==0.3.7
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
ipython==7.9.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
kiwisolver==1.1.0
kombu==4.6.5
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
networkx==2.4
numpy==1.17.3
packaging==19.2
parso==0.5.1
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pint==0.8.1
pkginfo==1.5.0.1
pluggy==0.13.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.3
PyYAML==5.1.2
readme-renderer==24.0
redis==3.3.11
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
snowballstemmer==2.0.0
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.2
toml==0.10.0
tox==3.14.0
tqdm==4.36.1
traitlets==4.3.3
twine==1.14.0
urllib3==1.25.6
vine==1.3.0
virtualenv==16.7.7
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
zeep==3.4.0
zipp==0.6.0
